+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pagination in the API
 - Pagination of the results in the webapp
 - Add a special display for empty results
+- Permalink to coredump and executable searchs
 ### Changed
 - Design improvements:
 	- Change the background color of hovered result rows